Lazy load EPG view on SectionCompositeViewController (#90)
Currently, opening SectionCompositeViewControllers with EPG enabled will cause
the whole React Native runtime to load, even if user never taps the EPG button.
Those changes modify this behavior to load EPG not when the screen is created
but rather when user opens the EPG view. It saves a lot of memory when swiping
through the home screen, because React Native is loaded only when it's used.